Osiris/flashblocks cleanup (#265)
* fix: block executor + e2e test with receipts root assertion

* chore: cleanup + add doc comments
