+++ crates/world/flashblocks/src/builder/mod.rs
@@ -3,7 +3,7 @@ use crate::{
         executor::{FlashblocksBlockBuilder, FlashblocksBlockExecutor},
         payload_txns::BestPayloadTxns,
     },
-    payload_builder_ctx::{PayloadBuilderCtx, PayloadBuilderCtxBuilder},
+    context::{PayloadBuilderCtx, PayloadBuilderCtxBuilder},
 };
 
 use alloy_consensus::BlockHeader;
