+++ crates/world/flashblocks/src/lib.rs
@@ -2,5 +2,5 @@
 #![allow(clippy::type_complexity)]
 pub mod args;
 pub mod builder;
-pub mod payload_builder_ctx;
+pub mod context;
 pub mod ws;
