--- crates/world/flashblocks/src/payload_builder_ctx/mod.rs
@@ -1,5 +1,0 @@
-mod builder;
-mod ctx;
-
-pub use builder::PayloadBuilderCtxBuilder;
-pub use ctx::PayloadBuilderCtx;
