+++ crates/world/payload/src/lib.rs
@@ -1,4 +1,4 @@
 #![cfg_attr(not(test), warn(unused_crate_dependencies))]
 pub mod builder;
 
-pub mod ctx;
+pub mod context;
